requirements.txt
# 基础科学计算
numpy>=1.24.0
pandas>=1.5.3
scipy>=1.11

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 机器学习
scikit-learn>=1.2.0

# 时间序列处理
tsfresh>=0.20
statsmodels>=0.14

# 工具包
pyyaml>=6.0
jupyter>=1.0.0

# 大数据处理
dask>=2023.3.0
pyarrow>=14.0.1
fastparquet>=2023.7.0

# 效率工具
pytz>=2023.3
tqdm>=4.65.0
python-dateutil>=2.8.2
psutil>=5.9.0

# 专用数据结构
intervaltree>=3.1.0

# 深度学习
torch>=2.0.0